requirements.txt
flask
requests
openpyxl
gunicorn
python-escpos
flask-cors
pyusb
